Splits reducer into multiple reducers
