+++ components/brave_rewards/ui/BUILD.gn
@@ -14,7 +14,9 @@ transpile_web_ui("ui") {
     "components/settingsPage.tsx",
     "constants/rewards_types.ts",
     "reducers/index.ts",
+    "reducers/grant_reducer.ts",
     "reducers/rewards_reducer.ts",
+    "reducers/wallet_reducer.ts",
     "storage.ts",
     "store.ts",
     "utils.ts",
